requirements2.txt
aiohttp==3.8.1
aiohttp-retry==2.4.6
aiosignal==1.2.0
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
asyncssh==2.8.1
atpublic==3.0.1
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
black==21.12b0
bleach==4.1.0
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cryptography==36.0.1
datasets==1.17.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
dill==0.3.4
diskcache==5.4.0
distro==1.6.0
dpath==2.0.5
dulwich==0.20.30
dvc==2.9.3
entrypoints==0.3
executing==0.8.2
filelock==3.4.2
flatten-dict==0.4.2
flufl.lock==7.0
frozenlist==1.2.0
fsspec==2022.1.0
ftfy==6.0.3
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.26
grandalf==0.6
huggingface-hub==0.4.0
idna==3.3
importlib-resources==5.4.0
ipykernel==6.7.0
ipython==8.0.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.1
jupyter-core==4.9.1
jupyter-server==1.13.3
jupyterlab==3.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.2
mailchecker==4.1.8
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==5.2.0
multiprocess==0.70.12.2
mypy-extensions==0.4.3
nanotime==0.5.2
nbclassic==0.3.5
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
nltk==3.5
notebook==6.4.7
numpy==1.22.1
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
phonenumbers==8.12.41
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
ply==3.11
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.1
pyarrow==6.0.1
pyasn1==0.4.8
pycparser==2.21
pydot==1.4.2
pygit2==1.7.2
Pygments==2.11.2
pygtrie==2.4.2
pyparsing==3.0.6
pyrsistent==0.18.1
python-benedict==0.24.3
python-dateutil==2.8.2
python-fsutil==0.5.0
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.10
requests==2.27.1
rich==11.0.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sacremoses==0.0.47
scikit-learn==0.23.2
scipy==1.7.3
scmrepo==0.0.7
Send2Trash==1.8.0
shortuuid==1.0.8
shtab==1.5.3
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
stack-data==0.1.4
tabulate==0.8.9
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
tomli==1.2.3
torch==1.9.0+cu111
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.15.0
typing-extensions==4.0.1
urllib3==1.26.8
voluptuous==0.12.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
widgetsnbextension==3.5.2
xmltodict==0.12.0
xxhash==2.0.2
yarl==1.7.2
zc.lockfile==2.0
zipp==3.7.0
